Update button for style
