+++ include/ui_panels/buttallfoos.h
@@ -20,14 +20,4 @@ void load_content(void *data);
 void open_options(void *data);
 void open_cmds(void *data);
 
-panel_t *create_start_button(program_t *p);
-panel_t *create_quit_button(program_t *p);
-panel_t *create_gquit_button(program_t *p);
-panel_t *create_continue_button(program_t *p);
-panel_t *create_inventory_button(program_t *p);
-panel_t *create_stats_button(program_t *p);
-panel_t *create_save_button(program_t *p);
-panel_t *create_load_button(program_t *p);
-panel_t *create_options_button(program_t *p);
-panel_t *create_cmds_button(program_t *p);
 #endif /* BUTTALLFOOS_H */
