+++ src/ui_panels/button_factory.c
@@ -11,18 +11,6 @@
 #include "theme.h"
 #include "tools.h"
 
-panel_t *make_butt(const char *label, void *trgt, button_foo_t foo, sfFont *font)
-{
-    sfVector2f size = {200, 50};
-    rectransform_t *sb_rect = rtrans_create_flexelem(size);
-    panel_t *butt = panel_text_button_create(
-        sb_rect, strdup(label), font, PTYPE_EVBUTT);
-
-    ((panel_button_t *)butt->data)->trgt = trgt;
-    butt_set_foos(butt, foo, NULL);
-    return butt;
-}
-
 static panel_button_t *button_create(void)
 {
     panel_button_t *bpanel = malloc(sizeof(panel_button_t));
@@ -41,10 +29,7 @@ static panel_button_t *button_create(void)
 }
 
 panel_t *panel_text_button_create(
-    rectransform_t *rect,
-    const char *str,
-    sfFont *font,
-    ptype_t type)
+    rectransform_t *rect, const char *str, sfFont *font, ptype_t type)
 {
     panel_button_t *bpanel = button_create();
     panel_t *panel = NULL;
@@ -62,9 +47,7 @@ panel_t *panel_text_button_create(
 }
 
 panel_t *panel_image_button_create(
-    rectransform_t *rect,
-    sfTexture *texture,
-    ptype_t type)
+    rectransform_t *rect, sfTexture *texture, ptype_t type)
 {
     panel_button_t *bpanel = button_create();
     panel_t *panel = NULL;
@@ -95,15 +78,4 @@ void panel_button_destroy(panel_t *panel)
         sfTexture_destroy(bpanel->texture);
     free(bpanel);
     panel_destroy(panel);
-}
-
-void butt_set_foos(panel_t *p, button_foo_t select, button_foo_t unselect)
-{
-    panel_button_t *bpanel = NULL;
-
-    if (!p)
-        return;
-    bpanel = (panel_button_t *)p->data;
-    bpanel->on_select = select;
-    bpanel->on_unselect = unselect;
 }
